+++ src/types/metaTypeCreator.test.ts
@@ -1,11 +1,6 @@
 import { MetaTypeCreator } from './metaTypeCreator'
 import { FirelordFirestore } from './firelordFirestore'
-import {
-	ErrorNullBanned,
-	ErrorInvalidDocumentOrCollectionID,
-	ErrorInvalidDocumentOrCollectionIDStart,
-	ErrorUnionInvolveObjectType,
-} from './error'
+import { ErrorNullBanned, ErrorUnionInvolveObjectType } from './error'
 import {
 	ArrayUnionOrRemove,
 	Increment,
@@ -17,57 +12,6 @@ import { NotTreatedAsObjectType } from './ref'
 import { IsTrue, IsSame, IsEqual } from './utils'
 
 describe('test Firelord type', () => {
-	it('test ID name', () => {
-		type A = MetaTypeCreator<
-			{
-				a: 1 | PossiblyReadAsUndefined
-				b:
-					| {
-							c: 'a' | PossiblyReadAsUndefined
-							d: { e: false } | PossiblyReadAsUndefined
-							f:
-								| {
-										g: Date | null | PossiblyReadAsUndefined
-										h: 2 | PossiblyReadAsUndefined
-								  }[]
-								| PossiblyReadAsUndefined
-					  }
-					| PossiblyReadAsUndefined
-			},
-			'/' | 'A/' | '~!@#$%^&*()_+-=' | '.',
-			'/' | '/A/' | 'A/A' | '.'
-		>
-
-		type CollectionIDTypeCheck<T extends string> = T[] extends (
-			| '~!@#$%^&*()_+-='
-			| ErrorInvalidDocumentOrCollectionIDStart<'Collection'>
-			| ErrorInvalidDocumentOrCollectionID<'Collection'>
-		)[]
-			? (
-					| '~!@#$%^&*()_+-='
-					| ErrorInvalidDocumentOrCollectionIDStart<'Collection'>
-					| ErrorInvalidDocumentOrCollectionID<'Collection'>
-			  )[] extends T[]
-				? true
-				: false
-			: false
-		type DocumentIDTypeCheck<T extends string> = T[] extends (
-			| ErrorInvalidDocumentOrCollectionIDStart<'Document'>
-			| ErrorInvalidDocumentOrCollectionID<'Document'>
-		)[]
-			? (
-					| ErrorInvalidDocumentOrCollectionIDStart<'Document'>
-					| ErrorInvalidDocumentOrCollectionID<'Document'>
-			  )[] extends T[]
-				? true
-				: false
-			: false
-
-		type collectionID = A['collectionID']
-		type docID = A['docID']
-		IsTrue<CollectionIDTypeCheck<collectionID>>()
-		IsTrue<DocumentIDTypeCheck<docID>>()
-	})
 	it('test read all as undefined', () => {
 		type A = MetaTypeCreator<
 			{
